+++ miamoremustachio/app/main.js
@@ -101,67 +101,4 @@ const toDo = {
   },
 };
 
-const fixtures = {
-  titles: {
-    valid: [ 'foo', 'bar', 'baz' ],
-    invalid: [ 1, null, 'a'.repeat(MAX_TITLE_LENGTH) ],
-  },
-  statuses: {
-    invalid: 'a',
-  },
-  ids: {
-    valid: [ 1, 2, 3 ],
-    invalid: [ '1', '2', '3' ],
-  },
-};
-
-const validTitles = [ ...fixtures.titles.valid ];
-const invalidTitles = [ ...fixtures.titles.invalid ];
-const invalidStatus = fixtures.statuses.invalid;
-const validIds = [ ...fixtures.ids.valid ];
-const invalidIds = [ ...fixtures.ids.invalid ];
-
-toDo.showTaskList();
-
-// TODO: add tests for tasks changing
-toDo.add(validTitles[0]);
-const testTask1 = toDo.list.find(task => task.title === validTitles[0]);
-console.log(testTask1.title === validTitles[0]); // ✓
-
-toDo.add(validTitles[1], DONE);
-const testTask2 = toDo.list.find(task => task.title === validTitles[1]);
-console.log(testTask2.status === DONE); // ✓
-
-toDo.add(invalidTitles[0], DONE);
-const testTask8 = toDo.list.find(task => task.title === invalidTitles[0]);
-console.log(testTask8 === undefined); // ✓
-
-toDo.add(invalidTitles[1], DONE);
-const testTask9 = toDo.list.find(task => task.title === invalidTitles[1]);
-console.log(testTask9 === undefined); // ✓
-
-toDo.add(invalidTitles[2], DONE);
-const testTask10 = toDo.list.find(task => task.title === invalidTitles[2]);
-console.log(testTask10 === undefined); // ✓
-
-toDo.add(validTitles[2], invalidStatus);
-const testTask3 = toDo.list.find(task => task.title === validTitles[2]);
-console.log(testTask3 === undefined); // ✓
-
-toDo.deleteByTitle(validTitles[0]);
-const testTask4 = toDo.list.find(task => task.title === validTitles[0]);
-console.log(testTask4 === undefined); // ✓
-
-toDo.deleteByTitle(validTitles[2]);
-const testTask5 = toDo.list.find(task => task.title === validTitles[2]);
-console.log(testTask5 === undefined); // ✓
-
-toDo.deleteById(validIds[0]);
-const testTask6 = toDo.list.find(task => task.id === validIds[0]);
-console.log(testTask6 === undefined); // ✓
-
-toDo.deleteById(invalidIds[1]);
-const testTask7 = toDo.list.find(task => task.id === validIds[1]);
-console.log(Boolean(testTask7)); // ✓
-
-toDo.showTaskList();
+module.exports = { toDo };
